UI update 1 (#14)
* change styling

* add different notification pages

* improve event UI

* rename functions

* change styling and add filters

* fake event creator creates for different blockchains

* improve dropdown UI

* adjust styling

* add chain filter

* fix styling

* fix clippy errors

* fix filters

* refactor

* fix sidebar button layout

+++ client/src/services/auth.rs
@@ -15,8 +15,8 @@ use crate::services::grpc::{KeplrLoginRequest, LoginResponse, RefreshAccessToken
 
 #[derive(Debug, Clone, Eq, PartialEq)]
 enum Role {
-    User = 0 as isize,
-    Admin = 1 as isize,
+    User = 0_isize,
+    Admin = 1_isize,
 }
 
 impl<'de> Deserialize<'de> for Role {
